+++ startup/Gaffer/tweakPlugCompatibility.py
@@ -43,8 +43,8 @@ def initSupportingOldSerialisation( self, *args, **kwargs ) :
 	if len( args ) == 1 and isinstance( args[0], str ) :
 
 		# Constructor calls that provide only a name but no value plug or
-		# data need to be rerouted. These calls occur when
-		# deserialising old scripts.
+		# data need to be rerouted. These calls occur when deserialising
+		# scripts saved in Gaffer 0.53 and earlier.
 
 		originalInit( self, None, args[0], **kwargs )